Mock npm executable to save a bunch of time in plugin install test

+++ tests/packaging/test_plugin_install.sh
@@ -6,6 +6,10 @@ unset PYTHONPATH
 
 source "${virtualenv_activate}"
 
+# Mock npm executable to save time
+export PATH="$(pwd)/mockpath:$PATH"
+rm -f npm_called.txt
+
 girder-install plugin "${source_path}/plugins/thumbnails"
 
 if [ $? -eq 0 ] ; then
@@ -13,7 +17,19 @@ if [ $? -eq 0 ] ; then
     exit 1
 fi
 
+if [ -f "npm_called.txt" ] ; then
+    echo "Npm should not have been called in failure case"
+    exit 1
+fi
+
 girder-install plugin -f "${source_path}/plugins/thumbnails" "${source_path}/plugins/oauth" || exit 1
+
+if [ ! -f "npm_called.txt" ] ; then
+    echo "Error: npm was not called during plugin install"
+    exit 1
+fi
+
+rm -f npm_called.txt
 
 dest_path=$(girder-install plugin-path)
 girder-install plugin -f "${dest_path}/thumbnails" || exit 1
@@ -23,4 +39,7 @@ if [ ! -f "${dest_path}/thumbnails/plugin.yml" ] ; then
     exit 1
 fi
 
-exit 0
+if [ ! -f "npm_called.txt" ] ; then
+    echo "Error: npm was not called during plugin install using plugin-path command"
+    exit 1
+fi
